+++ example/generate.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-pandoc -f markdown -t html --filter ./SatysfiFilter.sh -o test.html test.md
+pandoc -f markdown -t html --filter ./SatysfiFilter.sh -o generated/test.html test.md